make time stamps UINT64 instead of UINT32 to be compatible with DAQ

+++ icaruscode/CRT/CRTProducts/CRTData.hh
@@ -10,8 +10,8 @@ namespace crt {
 
       uint8_t  fMac5;
       uint32_t fEntry;
-      uint32_t fTs0;
-      uint32_t fTs1;
+      uint64_t fTs0;
+      uint64_t fTs1;
       uint16_t fAdc[64]; //only use indices 0-31 for C or M modules
 
       CRTData() {}
